# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRAMPO — Plataforma de matchmaking de carreiras com IA
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  cidade?: string | null;
  estado?: string | null;
  is_premium: boolean;
  premium_expires_at?: Date | null;
  created_at: Date;
}